backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
Pillow==9.5.0
python-dotenv==1.0.0
cloudinary==1.34.0
drf-yasg==1.21.7
django-cloudinary-storage==0.3.0
celery==5.3.6
redis==5.0.1
minio==7.2.0
django-celery-results==2.5.1
django-celery-beat==2.5.0 